+++ otterdog/models/repo_workflow_settings.py
@@ -54,6 +54,9 @@ class RepositoryWorkflowSettings(WorkflowSettings):
         if org_workflow_settings.default_workflow_permissions == "read":
             copy.default_workflow_permissions = UNSET  # type: ignore
 
+        if org_workflow_settings.actions_can_approve_pull_request_reviews is False:
+            copy.actions_can_approve_pull_request_reviews = UNSET  # type: ignore
+
         return copy
 
     def validate(self, context: ValidationContext, parent_object: Any) -> None:
@@ -76,11 +79,21 @@ class RepositoryWorkflowSettings(WorkflowSettings):
                 and self.default_workflow_permissions == "write"
             ):
                 context.add_failure(
-                    FailureType.ERROR,
+                    FailureType.WARNING,
                     f"{self.get_model_header(parent_object)} has 'default_workflow_permissions' of value "
                     f"'{self.default_workflow_permissions}', "
                     f"while on organization level it is restricted to "
-                    f"'{org_workflow_settings.default_workflow_permissions}'.",
+                    f"'{org_workflow_settings.default_workflow_permissions}', setting will be ignored.",
+                )
+
+            if (
+                org_workflow_settings.actions_can_approve_pull_request_reviews is False
+                and self.actions_can_approve_pull_request_reviews is True
+            ):
+                context.add_failure(
+                    FailureType.WARNING,
+                    f"{self.get_model_header(parent_object)} has 'actions_can_approve_pull_request_reviews' "
+                    "enabled, while on organization level it is disabled, setting will be ignored.",
                 )
 
     def include_field_for_diff_computation(self, field: dataclasses.Field) -> bool: